+++ lib/bitcoin_ui.dart
@@ -90,55 +90,65 @@ extension BitcoinTextStyle on TextStyle {
 
 // Buttons
 
-const defaultButtonWidth = 315.0;
 const defaultButtonHeight = 48.0;
 const defaultCornerRadius = 5.0;
 const defaultPadding = EdgeInsets.symmetric(vertical: 8, horizontal: 30);
-const defaultTintColor = Color(0xFFF89B2A);
-const defaultTextColor = Color(0xFFFFFFFF);
-const defaultDisabledTintColor = Color(0xFFF4F4F4);
-const defaultDisabledTextColor = Color(0xFFBBBBBB);
-const defaultDisabledOutlineColor = Color(0xFFDEDEDE);
 
 class BitcoinButtonFilled extends StatelessWidget {
   final String title;
   final TextStyle? textStyle;
-  final double width;
-  final double height;
-  final double cornerRadius;
-  final Color tintColor;
-  final Color textColor;
-  final Color disabledTintColor;
-  final Color disabledTextColor;
+  final double? width;
+  final double? height;
+  final double? cornerRadius;
+  final Color? tintColor;
+  final Color? textColor;
+  final Color? disabledTintColor;
+  final Color? disabledTextColor;
   final bool disabled;
   final bool isLoading;
+  final bool isCapsule;
   final VoidCallback? onPressed;
 
   const BitcoinButtonFilled({
     Key? key,
     required this.title,
     this.textStyle,
-    this.width = defaultButtonWidth,
-    this.height = defaultButtonHeight,
-    this.cornerRadius = defaultCornerRadius,
-    this.tintColor = defaultTintColor,
-    this.textColor = defaultTextColor,
-    this.disabledTintColor = defaultDisabledTintColor,
-    this.disabledTextColor = defaultDisabledTextColor,
+    this.width,
+    this.height,
+    this.cornerRadius,
+    this.tintColor,
+    this.textColor,
+    this.disabledTintColor,
+    this.disabledTextColor,
     this.disabled = false,
     this.isLoading = false,
+    this.isCapsule = true,
     required this.onPressed,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    // Set conditional button values
+    double screenWidth = MediaQuery.of(context).size.width;
+    double buttonWidth = width ?? screenWidth - 2 * 16;
+    double buttonHeight = height ?? defaultButtonHeight;
+    double buttonCornerRadius =
+        cornerRadius ?? (isCapsule ? buttonHeight / 2 : defaultCornerRadius);
+    Color buttonColor = disabled
+        ? disabledTintColor ?? Theme.of(context).colorScheme.secondary
+        : tintColor ?? Theme.of(context).colorScheme.primary;
+    Color buttonTextColor = disabled
+        ? disabledTextColor ?? Theme.of(context).colorScheme.onSecondary
+        : textColor ?? Theme.of(context).colorScheme.onPrimary;
+
     if (Platform.isIOS) {
       return SizedBox(
-          width: width,
+          width: buttonWidth,
           height: height,
           child: CupertinoButton(
-              color: disabled ? disabledTintColor : tintColor,
-              borderRadius: BorderRadius.all(Radius.circular(cornerRadius)),
+              color: buttonColor,
+              borderRadius:
+                  BorderRadius.all(Radius.circular(buttonCornerRadius)),
               padding: defaultPadding,
               onPressed: disabled
                   ? null
@@ -147,25 +157,29 @@ class BitcoinButtonFilled extends StatelessWidget {
                       : onPressed,
               child: isLoading
                   ? SizedBox(
-                      height: height * 0.5,
-                      width: height * 0.5,
+                      height: buttonHeight * 0.5,
+                      width: buttonHeight * 0.5,
                       child: Center(
-                          child: CupertinoActivityIndicator(color: textColor)),
+                          child: CupertinoActivityIndicator(
+                              color: buttonTextColor)),
                     )
                   : Text(title,
                       style: textStyle ??
-                          BitcoinTextStyle.title5(
-                              disabled ? disabledTextColor : textColor))));
+                          Theme.of(context)
+                              .textTheme
+                              .titleSmall
+                              ?.copyWith(color: buttonTextColor))));
     } else {
       return ElevatedButton(
           style: ButtonStyle(
-            backgroundColor: MaterialStateProperty.all(
-                disabled ? disabledTintColor : tintColor),
+            elevation: const MaterialStatePropertyAll(0),
+            backgroundColor: MaterialStateProperty.all(buttonColor),
             padding: MaterialStateProperty.all(defaultPadding),
-            fixedSize: MaterialStatePropertyAll(Size(width, height)),
+            fixedSize:
+                MaterialStatePropertyAll(Size(buttonWidth, buttonHeight)),
             shape: MaterialStateProperty.all(
               RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(cornerRadius),
+                borderRadius: BorderRadius.circular(buttonCornerRadius),
               ),
             ),
             enableFeedback: true,
@@ -177,17 +191,19 @@ class BitcoinButtonFilled extends StatelessWidget {
                   : onPressed,
           child: isLoading
               ? SizedBox(
-                  height: height * 0.5,
-                  width: height * 0.5,
+                  height: buttonHeight * 0.5,
+                  width: buttonHeight * 0.5,
                   child: Center(
                       child: CircularProgressIndicator(
-                          color: textColor, strokeWidth: 2)),
+                          color: buttonTextColor, strokeWidth: 2)),
                 )
               : Text(
                   title,
                   style: textStyle ??
-                      BitcoinTextStyle.title5(
-                          disabled ? disabledTextColor : textColor),
+                      Theme.of(context)
+                          .textTheme
+                          .titleSmall
+                          ?.copyWith(color: buttonTextColor),
                 ));
     }
   }
@@ -196,48 +212,60 @@ class BitcoinButtonFilled extends StatelessWidget {
 class BitcoinButtonOutlined extends StatelessWidget {
   final String title;
   final TextStyle? textStyle;
-  final double width;
-  final double height;
-  final double cornerRadius;
-  final Color tintColor;
-  final Color disabledTintColor;
+  final double? width;
+  final double? height;
+  final double? cornerRadius;
+  final Color? tintColor;
+  final Color? disabledTintColor;
   final bool disabled;
   final bool isLoading;
+  final bool isCapsule;
   final VoidCallback? onPressed;
 
   const BitcoinButtonOutlined({
     Key? key,
     required this.title,
     this.textStyle,
-    this.width = defaultButtonWidth,
-    this.height = defaultButtonHeight,
-    this.cornerRadius = defaultCornerRadius,
-    this.tintColor = defaultTintColor,
-    this.disabledTintColor = defaultDisabledTintColor,
+    this.width,
+    this.height,
+    this.cornerRadius,
+    this.tintColor,
+    this.disabledTintColor,
     this.disabled = false,
     this.isLoading = false,
+    this.isCapsule = true,
     required this.onPressed,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    // Set conditional button values
+    double screenWidth = MediaQuery.of(context).size.width;
+    double buttonWidth = width ?? screenWidth - 2 * 16;
+    double buttonHeight = height ?? defaultButtonHeight;
+    double buttonCornerRadius =
+        cornerRadius ?? (isCapsule ? buttonHeight / 2 : defaultCornerRadius);
+    Color buttonColor = disabled
+        ? disabledTintColor ?? Theme.of(context).colorScheme.secondary
+        : tintColor ?? Theme.of(context).colorScheme.primary;
+
     if (Platform.isIOS) {
       return SizedBox(
-          width: width,
-          height: height,
+          width: buttonWidth,
+          height: buttonHeight,
           child: OutlinedButton(
               // Can't use CupertinoButton here as it has no ability to set outlines
               style: ButtonStyle(
                 elevation: const MaterialStatePropertyAll(0),
                 backgroundColor: MaterialStateProperty.all(Colors.transparent),
                 padding: MaterialStateProperty.all(defaultPadding),
-                fixedSize: MaterialStatePropertyAll(Size(width, height)),
-                side: MaterialStatePropertyAll(BorderSide(
-                    width: 2.0,
-                    color: disabled ? disabledTintColor : tintColor)),
+                fixedSize:
+                    MaterialStatePropertyAll(Size(buttonWidth, buttonHeight)),
+                side: MaterialStatePropertyAll(
+                    BorderSide(width: 2.0, color: buttonColor)),
                 shape: MaterialStateProperty.all(
                   RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(cornerRadius),
+                    borderRadius: BorderRadius.circular(buttonCornerRadius),
                   ),
                 ),
                 enableFeedback: true,
@@ -249,28 +277,31 @@ class BitcoinButtonOutlined extends StatelessWidget {
                       : onPressed,
               child: isLoading
                   ? SizedBox(
-                      height: height * 0.5,
-                      width: height * 0.5,
+                      height: buttonHeight * 0.5,
+                      width: buttonHeight * 0.5,
                       child: Center(
-                          child: CupertinoActivityIndicator(
-                              color: disabled ? disabledTintColor : tintColor)),
+                          child:
+                              CupertinoActivityIndicator(color: buttonColor)),
                     )
                   : Text(title,
                       style: textStyle ??
-                          BitcoinTextStyle.title5(
-                              disabled ? disabledTintColor : tintColor))));
+                          Theme.of(context)
+                              .textTheme
+                              .titleSmall
+                              ?.copyWith(color: buttonColor))));
     } else {
       return ElevatedButton(
           style: ButtonStyle(
             elevation: const MaterialStatePropertyAll(0),
             backgroundColor: MaterialStateProperty.all(Colors.transparent),
             padding: MaterialStateProperty.all(defaultPadding),
-            fixedSize: MaterialStatePropertyAll(Size(width, height)),
-            side: MaterialStatePropertyAll(BorderSide(
-                width: 2.0, color: disabled ? disabledTintColor : tintColor)),
+            fixedSize:
+                MaterialStatePropertyAll(Size(buttonWidth, buttonHeight)),
+            side: MaterialStatePropertyAll(
+                BorderSide(width: 2.0, color: buttonColor)),
             shape: MaterialStateProperty.all(
               RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(cornerRadius),
+                borderRadius: BorderRadius.circular(buttonCornerRadius),
               ),
             ),
             enableFeedback: true,
@@ -282,8 +313,8 @@ class BitcoinButtonOutlined extends StatelessWidget {
                   : onPressed,
           child: isLoading
               ? SizedBox(
-                  height: height * 0.5,
-                  width: height * 0.5,
+                  height: buttonHeight * 0.5,
+                  width: buttonHeight * 0.5,
                   child: Center(
                       child: CircularProgressIndicator(
                           color: disabled ? disabledTintColor : tintColor,
@@ -292,8 +323,10 @@ class BitcoinButtonOutlined extends StatelessWidget {
               : Text(
                   title,
                   style: textStyle ??
-                      BitcoinTextStyle.title5(
-                          disabled ? disabledTintColor : tintColor),
+                      Theme.of(context)
+                          .textTheme
+                          .titleSmall
+                          ?.copyWith(color: buttonColor),
                 ));
     }
   }
@@ -302,38 +335,51 @@ class BitcoinButtonOutlined extends StatelessWidget {
 class BitcoinButtonPlain extends StatelessWidget {
   final String title;
   final TextStyle? textStyle;
-  final double width;
-  final double height;
-  final double cornerRadius;
-  final Color tintColor;
-  final Color disabledTintColor;
+  final double? width;
+  final double? height;
+  final double? cornerRadius;
+  final Color? tintColor;
+  final Color? disabledTintColor;
   final bool disabled;
   final bool isLoading;
+  final bool isCapsule;
   final VoidCallback? onPressed;
 
   const BitcoinButtonPlain({
     Key? key,
     required this.title,
     this.textStyle,
-    this.width = defaultButtonWidth,
-    this.height = defaultButtonHeight,
-    this.cornerRadius = defaultCornerRadius,
-    this.tintColor = defaultTintColor,
-    this.disabledTintColor = defaultDisabledTintColor,
+    this.width,
+    this.height,
+    this.cornerRadius,
+    this.tintColor,
+    this.disabledTintColor,
     this.disabled = false,
     this.isLoading = false,
+    this.isCapsule = true,
     required this.onPressed,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    // Set conditional button values
+    double screenWidth = MediaQuery.of(context).size.width;
+    double buttonWidth = width ?? screenWidth - 2 * 16;
+    double buttonHeight = height ?? defaultButtonHeight;
+    double buttonCornerRadius =
+        cornerRadius ?? (isCapsule ? buttonHeight / 2 : defaultCornerRadius);
+    Color buttonColor = disabled
+        ? disabledTintColor ?? Theme.of(context).colorScheme.secondary
+        : tintColor ?? Theme.of(context).colorScheme.primary;
+
     if (Platform.isIOS) {
       return SizedBox(
-          width: width,
-          height: height,
+          width: buttonWidth,
+          height: buttonHeight,
           child: CupertinoButton(
               color: Colors.transparent,
-              borderRadius: BorderRadius.all(Radius.circular(cornerRadius)),
+              borderRadius:
+                  BorderRadius.all(Radius.circular(buttonCornerRadius)),
               padding: defaultPadding,
               onPressed: disabled
                   ? null
@@ -342,26 +388,29 @@ class BitcoinButtonPlain extends StatelessWidget {
                       : onPressed,
               child: isLoading
                   ? SizedBox(
-                      height: height * 0.5,
-                      width: height * 0.5,
+                      height: buttonHeight * 0.5,
+                      width: buttonHeight * 0.5,
                       child: Center(
-                          child: CupertinoActivityIndicator(
-                              color: disabled ? disabledTintColor : tintColor)),
+                          child:
+                              CupertinoActivityIndicator(color: buttonColor)),
                     )
                   : Text(title,
                       style: textStyle ??
-                          BitcoinTextStyle.title5(
-                              disabled ? disabledTintColor : tintColor))));
+                          Theme.of(context)
+                              .textTheme
+                              .titleSmall
+                              ?.copyWith(color: buttonColor))));
     } else {
       return ElevatedButton(
           style: ButtonStyle(
             elevation: const MaterialStatePropertyAll(0),
             backgroundColor: MaterialStateProperty.all(Colors.transparent),
             padding: MaterialStateProperty.all(defaultPadding),
-            fixedSize: MaterialStatePropertyAll(Size(width, height)),
+            fixedSize:
+                MaterialStatePropertyAll(Size(buttonWidth, buttonHeight)),
             shape: MaterialStateProperty.all(
               RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(cornerRadius),
+                borderRadius: BorderRadius.circular(buttonCornerRadius),
               ),
             ),
             enableFeedback: true,
@@ -373,18 +422,19 @@ class BitcoinButtonPlain extends StatelessWidget {
                   : onPressed,
           child: isLoading
               ? SizedBox(
-                  height: height * 0.5,
-                  width: height * 0.5,
+                  height: buttonHeight * 0.5,
+                  width: buttonHeight * 0.5,
                   child: Center(
                       child: CircularProgressIndicator(
-                          color: disabled ? disabledTintColor : tintColor,
-                          strokeWidth: 2)),
+                          color: buttonColor, strokeWidth: 2)),
                 )
               : Text(
                   title,
                   style: textStyle ??
-                      BitcoinTextStyle.title5(
-                          disabled ? disabledTintColor : tintColor),
+                      Theme.of(context)
+                          .textTheme
+                          .titleSmall
+                          ?.copyWith(color: buttonColor),
                 ));
     }
   }